+++ app/EntityType.php
@@ -34,15 +34,21 @@ class EntityType extends Model
             ->logOnlyDirty();
     }
 
-    public function setRelationInfo($color, $isRoot = false, $subTypes = []) {
-        $this->is_root = $isRoot;
-        $this->color = $color;
-        EntityTypeRelation::where('parent_id', $this->id)->delete();
-        foreach($subTypes as $type) {
-            $relation = new EntityTypeRelation();
-            $relation->parent_id = $this->id;
-            $relation->child_id = $type;
-            $relation->save();
+    public function setRelationInfo($data) {
+        if(array_key_exists('is_root', $data)) {
+            $this->is_root = $data['is_root'];
+        }
+        if(array_key_exists('color', $data)) {
+            $this->color = $data['color'];
+        }
+        if(array_key_exists('sub_entity_types', $data)) {
+            EntityTypeRelation::where('parent_id', $this->id)->delete();
+            foreach($data['sub_entity_types'] as $type) {
+                $relation = new EntityTypeRelation();
+                $relation->parent_id = $this->id;
+                $relation->child_id = $type;
+                $relation->save();
+            }
         }
         $this->save();
     }