+++ resources/js/components/tree/TreeMenu.vue
@@ -26,6 +26,32 @@
             <a
                 class="dropdown-item"
                 href="#"
+                @click.stop.prevent="addEntity('above')"
+                @dblclick.stop.prevent=""
+            >
+                <i class="fas fa-fw fa-turn-up text-success" />
+                <span class="ms-2">
+                    {{ t('main.entity.tree.contextmenu.add_above') }}
+                </span>
+            </a>
+        </li>
+        <li>
+            <a
+                class="dropdown-item"
+                href="#"
+                @click.stop.prevent="addEntity('below')"
+                @dblclick.stop.prevent=""
+            >
+                <i class="fas fa-fw fa-turn-down text-success" />
+                <span class="ms-2">
+                    {{ t('main.entity.tree.contextmenu.add_below') }}
+                </span>
+            </a>
+        </li>
+        <li>
+            <a
+                class="dropdown-item"
+                href="#"
                 @click.stop.prevent="duplicateEntity"
                 @dblclick.stop.prevent=""
             >
@@ -109,8 +135,16 @@
                 context.emit('close');
             });
 
-            const addEntity = _ => {
-                showAddEntity(props.data);
+            const addEntity = where => {
+                if(where == 'above') {
+                    const parent = entityStore.getEntity(props.data.root_entity_id);
+                    showAddEntity(parent, null, props.data.rank);
+                } else if(where == 'below') {
+                    const parent = entityStore.getEntity(props.data.root_entity_id);
+                    showAddEntity(parent, null, props.data.rank + 1);
+                } else {
+                    showAddEntity(props.data);
+                }
             };
             const duplicateEntity = _ => {
                 duplicateEntityApi(props.data).then(data => {